fix: separate createWritable polyfill
